Develop lsy0

+++ src/main/java/com/herejava/book/controller/BookViewServlet.java
@@ -32,7 +32,7 @@ public class BookViewServlet extends HttpServlet {
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		request.setCharacterEncoding("utf-8");
-		int bookNo = Integer.parseInt(request.getParameter("bookNo"));
+		long bookNo = Long.parseLong(request.getParameter("bookNo"));
 		BookService service = new BookService();
 		Book b = service.selectOneBook(bookNo);
 		RequestDispatcher view = request.getRequestDispatcher("/WEB-INF/views/book/bookView.jsp");
